enable CPU row-wise sharding (#1396)
Summary:

Add a new flag to fused parameters to enforce sharding happens on CPU, and also handle BC.

Current in-memory publish for hstu model is on CPU. DI publish without GMPP needs to shard in CPU memory and do TGIF transformations on CPU as well. We did the same thing for sharded tensor pool to enable Prospector model's multi-card D44244973.

I tried to not use any additional parameter, and just assign dist_input and lookups to CPU when self.device is CPU, but it breaks internal tests who did dmp_modol.to("cuda") or dmp_model.copy("cuda"),  because dmp.to() and dmp.copy doesn't move to cuda device for sharded modules. So I introduced a new parameter in fused params to avoid BC breakage.

Reviewed By: YazhiGao

Differential Revision: D49256608