Revert D49256608: Multisect successfully blamed "D49256608: [TorchRec] enable CPU row-wise sharding" for test or build failures (#1397)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1397

This diff is reverting D49256608
D49256608: [TorchRec] enable CPU row-wise sharding by jiayisuse has been identified to be causing the following test or build failures:

Tests affected:
- [torchrec/github/examples/retrieval/tests:test_two_tower_train - torchrec.github.examples.retrieval.tests.test_two_tower_train.TrainTest: test_train_function](https://www.internalfb.com/intern/test/281475048734207/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3072404
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: IvanKobzarev

Differential Revision: D49361458

fbshipit-source-id: 38f49a14a5e3ea63bc9daea4e9e8f3b622af01cd