Add dedup flag to EC sharder as option.
Summary:
Passing as non-global flag will allow us to plan around this feature, and avoid a couple of sharp edge cases such a if a user first shards a model, and then sets the flag. It could be easy to think that they enabled this feature, but it doesn't.

Will remove global flag in a later diff.

Reviewed By: dstaay-fb

Differential Revision: D48607324