Use hbm bw for uvm caching when we are using prefetching
Summary:
Choose to use hbm perf when we are using prefetch_pipeline.

Underlying assumption is that we have correct per table cache_load_factor when using prefetch_pipeline.

Re the underlying assumption, there is no way the planner can tell if the cache_load_factor is good or not, since it doesn't have access to the index distribution, and pooling_factor alone is not enough.

Differential Revision: D49084591