Correct ZCH shard metadata (#1440)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1440

This change modify how we persist shard metadata for MC modules over ZCH table, which is basically the pair for (offset, size, sum).

In the past, we use local ZCH shard size and rank id to calculate metadata. This is only right if all shard sizes are the same (embedding table size is divisible by world_size). And bug happens when it's not the case : P850632086

The best way of computation is to use local shards metadata alone to calculate ZCH shard metadata. This is not possible unfortunately for now, since ZCH shard is basically table shard * ZCH ratio, which lose precision in the calculation. So sum(shard * ZCH_ratio) != sum(shard) * ZCH_ratio.

Discussed with YLGH we decide to use PG to communicate the right ZCH table size per rank. This does incur extra communication but is fine since it only happens on initialization

Reviewed By: duduyi2013, YLGH

Differential Revision: D50112671

fbshipit-source-id: 854ec6d76ab8c8c180e1dc0a156dd3090bdfb0f8