Allow Missing Labels for MTML models
Summary: Some MTML Looper models contain tasks that have little/no data, with no labels in a batch. This will lead to NaN NE and result in the model failing when it shouldn't. This diff allows the BinaryNE metric to allow for no samples in a batch, normally producing nan values and instead producing an epsilon NE.

Reviewed By: zainhuda

Differential Revision: D50273422

fbshipit-source-id: 0e4b7ae168c038f97c442918f1df0e092ba9088d