Add option to balance modules to partitioners (#1509)
Summary:

This diff will make sure module count on GPUs are balanced. The way we do it is to sort the tables from a smaller module first.

This should work fine with dependency (aka towers), since dependency is either child_path or starts with child_path, which is the path.

Reviewed By: bigning, richqyz

Differential Revision: D51180772