Add option to balance modules to partitioners (#1509)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1509

This diff will make sure module count on GPUs are balanced. The way we do it is to sort the tables from a smaller module first.

This should work fine with dependency (aka towers), since dependency is either child_path or starts with child_path, which is the path.

Reviewed By: bigning, richqyz

Differential Revision: D51180772

fbshipit-source-id: a31e2e3fb341ab86622a6f365e9b0e192703099a