Enhance KeyedOptimizer's `state_dict` and `load_state_dict` to handle nested state dictionaries cleanly (#1533)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1533

Pair programmed with Tsung-Hsien Lee (zong).

In order to support Distributed Shampoo (v2) checkpointing, we need to enhance `KeyedOptimizer`'s `state_dict` and `load_state_dict` functionality to support nested dictionaries, where the leaf node of any nested dictionary could be an `OptimizerModule` or `Tensor`. To do this, we create a recursive function for saving and updating a state dictionary in lieu of assuming a single-layer dictionary.

We also modify tests within `test_keyed.py` in order to test this new functionality.

Note that we must add the case for `ShardedTensor` in the recursive function `_update_param_state_dict_object`. (Should this be the case?)

Reviewed By: YLGH

Differential Revision: D50620406

fbshipit-source-id: 009a6964c63bafc863e30449e1ab603b7aac5425