fix fbgemm backward compat for PermutePooledEmbeddings (#1534)
Summary:

We modified `PermutePooledEmbeddings` in fbgemm directory, but due to packaging it will still be using the original version. The original version is a module that has registered buffers that will show up in the state dict causing incompatibility. This diff moves the new non state dict version inside torchrec to avoid the packaging issue.

Differential Revision: D51500998