Fix NDCG Metric on GPU
Summary: Previously, using the NDCG metric on GPU failed in f506598063 with the error: Expected all tensors to be on the same device, but found at least two devices. This diff ensures that the get_ndcg_states function returns tensors that are on the correct device that the model is being trained on.

Differential Revision: D51603318