Fix NDCG Metric on GPU (#1543)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1543

Previously, using the NDCG metric on GPU failed in f506598063 with the error: Expected all tensors to be on the same device, but found at least two devices. This diff ensures that the get_ndcg_states function returns tensors that are on the correct device that the model is being trained on.

Differential Revision: D51603318

fbshipit-source-id: 9c139a3136ea8c8113f90187f2d4dc2e9a092128