MCM Fix for ig integration (#1541)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1541

Some bug fixes during the integration test in PyPER O3:
# fix #1
_embedding_bag_collection (ShardedEmbeddingBagCollection) is not really called by input_dist (because the same thing is already distributed by ShardedManagedCollisionCollection) . So it never get a chance to initiate _input_dist. As a result, TREC pipelining thinks it's not ready for input distribution.
This is not expected, since the module is not used in the stage anyway, nor should it be put in fused a2a communication. With this change, https://fburl.com/code/ud8lnixv it'll satisfy the assertion, meanwhile doesn't carry _input_dists so won't be put into fused a2a.
# fix #2
ManagedCollisionCollection.forward is not traceable because it uses unwarpped KeyedJaggedTensor.from_jt_dict. We don't care about its internal detail so just keep it atomic.
# fix #3
Due to how remap table is set, MCHManagedCollisionModule doesn't support i32 id list for now. An easy fix is to convert to i64 regardless. A more memory efficient fix is probably change the remapper to i32 if necessary

Reviewed By: dstaay-fb

Differential Revision: D51601041

fbshipit-source-id: 95cf346b5247f1d5afb6643ecfd7dca4b3c4d575