fix publish error from fx tracing check in unsharded ebc (#1546)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1546

The reindexing logic in the ebc module isn't compatible with fx tracing, due to the KJT containing proxies. To circumvent this we skip this logic with a `is_fx_tracing()` function check.

During mvai's publish flow this introduces a jit error:
```
RuntimeError:
Could not export Python function call 'is_fx_tracing'. Remove calls to Python functions before export. Did you forget to add script or script_method annotation? If this is a nn.ModuleList, add it to __constants__:
  File "/mnt/xarfuse/uid-29063/e95415f6-seed-nspid4026534553_cgpid126435444-ns-4026534550/torchrec/modules/embedding_modules.py", line 203
                ).float()
                pooled_embeddings.append(res)
        if features.inverse_indices_or_none() is not None and not is_fx_tracing():
                                                                  ~~~~~~~~~~~~~ <--- HERE
            inverse_indices = []
            index_per_name = {
```

to fix this we fx wrap the conditional reindexing logic. Due to LRM split logic, however, we cannot pass KJT into the fx wrap reindexing function after lookup because that will tag the node as local, which is same as KJT. This creates an extra local split because logic after the lookup is supposed to be merge. To fix this we only optionally pass the inverse indices as a list of tensors after the lookup, because tensors will not cause retagging in node and we added support so that None would not as well.

Reviewed By: 842974287

Differential Revision: D51718962

fbshipit-source-id: d8dc3e493a65bf76ecba1143d72fde060864015b