reworked: move tensor concatenation to compute step from update step (#1544)
Summary:

We don't need to concatenate the tensor on every update step, since it is an expensive operation (creates a new tensor and allocates new memory every call, as tensors are contiguous) we can call tensor.concat on the compute step instead. Which happens every compute_interval_step batches. This optimization should boost performance of models using AUC with no regression in metric quality.

An earlier iteration of this was landed (D51176437) and caused a SEV as it appended model output to state indefinitely causing OOMs. This is mitigated in this version by strictly adhering to `self._window_size` checking on every update and evicting old tensors as soon as they exceed this limit.

Reviewed By: dstaay-fb

Differential Revision: D51399384