3.3x speed up for partitioner for large models
Summary:
Optimize partitioner so device_partition doesn't resort the entire
devices list on placement of every shard.

D45372819 improved this 21% by keeping a shallow mostly sorted list
between sorts to reduce the amount of reordering the sort needed.

This diffs removes the sort completely, instead placing the devices in a
minheap data structure that can generate the best candidate by perf in
O(log(n)) vs O(nlog(n)) for the sort. Because of correlation between
perf and storage, devices with low perf also tend to have more storage
available, so we frequently don't need to test more than a handful of
candidates before we find one with available storage.

This gives a ~3.3x saving for a real model with 5,384 tables on 1024 GPUs.

Between this diff and D51490378, the planner is now ~6x faster for
large models.

Reviewed By: bigning, henrylhtsang

Differential Revision: D51567156