Revert D51399384: Multisect successfully blamed "D51399384: [torchrec][recmetrics] reworked: move tensor concatenation to compute step from update step" for otest failure
Summary:
This diff is reverting D51399384
D51399384: [torchrec][recmetrics] reworked: move tensor concatenation to compute step from update step by zainhuda has been identified to be causing the following test failure:

Tests affected:
- [aps_models/exploration/ig/tests:e2e_esr_multi_table_train_local_gpu_test - aps_models.exploration.ig.tests.e2e_esr_multi_table_train_local_gpu_test.TestE2EESRTrainLocalGPU: test_local_multi_process_training](https://www.internalfb.com/intern/test/281475099588398/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3785687
Here are the tasks that are relevant to this breakage:


We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Differential Revision: D52164472