improve AUC memory usage test to include multiple updates and reset (#1635)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1635

Metamate:

This diff improves the AUC memory usage test in the `torchrec/metrics/tests/test_metric_module.py` file. The changes include adding a test for multiple updates and resets to ensure that memory usage does not increase unexpectedly. The code change includes adding a loop to update the metric module multiple times and checking that the memory usage does not increase beyond the expected amount. The test is designed to catch any unexpected memory usage increases and ensure that the metric module is working as expected.

Reviewed By: leitian, henrylhtsang

Differential Revision: D52838208

fbshipit-source-id: 359bca52db5d953eb633e567ad6e07c7bf2c68d6