Fail early if no sharding option found for table (#1657)
Summary:

Currently we raise an exception if no sharding options are found for the first table. If a sharding option is found for the first table, but not for the second table, no exception is raised.

This causes the error to be [raised later when sharding the model](https://github.com/pytorch/torchrec/blob/77974f229ce7e229664fbe199e1308cc37a91d7f/torchrec/distributed/embeddingbag.py#L217-L218), which is harder to debug.

Reviewed By: henrylhtsang

Differential Revision: D53044797