Allow passing feature names through ParameterConstraints and ShardingOption
Summary:
Allow users to pass in feature_names. Preferably, it should be same as table.feature_names.

In the case when no constraints are passed in, the feature_names will be None. 

Another option is to force extract feature names from BaseEmbeddingConfig, but that seems to be against the convention.

Differential Revision: D53454672