Allow passing feature names through ParameterConstraints and ShardingOption (#1686)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1686

Allow users to pass in feature_names. Preferably, it should be same as table.feature_names.

In the case when no constraints are passed in, the feature_names will be None.

Another option is to force extract feature names from BaseEmbeddingConfig, but that seems to be against the convention.

Reviewed By: ge0405

Differential Revision: D53454672

fbshipit-source-id: 112d48baa3c2072c4004beafc64312ffde119901