Small syntactic changes for dynamo compatibility
Summary:
Dynamo has some gaps in support of generators, list comprehension etc.
Avoiding them for now with syntactic changes

Previous diff was reverted because recat was created on the target device from the start.
Then with per-item manipulations it was writing directly to device (which broke freya training as it looks like freya does not support per-item changes).

In this diff recat is created on "cpu", the same as List[int] in original version.

Reviewed By: MatthewWEdwards

Differential Revision: D54192498

fbshipit-source-id: bdfffc5f207fa200c9b207969225c2bcbdf94e7a