Add staged train pipeline to torchrec and refactor train pipeline organization (#1624)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1624

# train pipeline code structure
- moves train_pipeline.py under /train_pipeline dir
- moves helper functions, classes to utils.py
- keeps imports under` torchrec.train_pipeline`

# staged train pipeline
Variable stage train pipeline that can support arbitrary number of stages. Note that this is a pre-forward scheduling pipeline, this means the forward is expected executed after the last stage in the pipeline and the last stage in the pipeline has to be the start SDD stage. This is because ShardedModule forward (overwritten by pipeline) has to consume the input from pipeline context post input dist.

The design is illustrated in the figure below. A pipeline is composed of K stages, each one depends on its precedence. Different stage may be executed in different or the same streams, with multiple batches concurrently executed in the same iteration. For example, in the image, batch[0] is the oldest batch that has passed H2D, preproc, and SDD stages, and will be running through comp in the current iteration. Batch[1], on the other hand, is the second oldest batch that will execute SDD in the current iteration. Similarly for other batches.

 {F1150156522}

with this, 4 batches will be handled together in the same iteration, while each of them is under a different stage. When an iteration is done, there will be a advance step to copy newer data (slots larger index) to older data (slots with smaller index) so that they will be handled by the next stage in the next iteration.

For SDD handling, we currently wrap on top of existing torchrec utilities. This part could potentially be improved in the future.

Some additional things on top of this:
* Adding fill callback for start sdd
* Adding await sdd as a callback
* Modifying progress to be walrus'able

Reviewed By: dracifer, sarckk

Differential Revision: D51182804