Support custom embedding config (#1806)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1806

Currently the torchrec benchmark takes  the DLRM config only. This diff tries to make it more flexible and it can take a json file with the following format:
{{table_name: {num_embeddings: , embedding_dim, pooling_factor (optional)}}

Pooling factor is optional. If we don't pass in pooling factor, it uses 10 for every table. If passed in, we'll generate a normal distribution for it to simulate the real situation.

Reviewed By: PaulZhang12

Differential Revision: D55055970

fbshipit-source-id: 0519e5ad0795d5a8e9011202c69d471892b376a8