Broadcast DP tables when init
Summary:
# Problem

Problem is once we wrap DP tables with DDP, the parameters are the same, but not synced. So if we reset the parameters in the table without using the same manual seed, it could cause the DP table parameters to be different.

In other words, DP tables would be initialized differently.

# Fix

There are a few ways to fix it. This is the way we believe to be least invasive and follow the spirit of the api.

What we do:
1. Broadcast DP tables of rank 0 to all other ranks
2. This only happens during init, or when we call reset_parameters

Reviewed By: joshuadeng

Differential Revision: D55227979

fbshipit-source-id: 52c225501df86cf35b196afff48dc2fa88e7a2c3