Fix broken test & increase the test count for test_embedding_sharding (#1920)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1920

As title suggest, fix the previously broken tests.

Also, we realized that the test count (10) is too small compared to the total enumerated conditions, it's very likely a test could be broken in some branch but not covered in several runs.

Increasing the total test sample by 10x will mitigate the issue

Reviewed By: henrylhtsang

Differential Revision: D56478625

fbshipit-source-id: d835859f70feaa11c0234a6ca788c45352e84ab4