Implement pipeline-aware shard estimator and turn it on by default for prefetch pipeline (#1924)
Summary:

Lots of rank load imbalance comes from underestimation of sparse, which is not only a function of parameter, but the HBM usage during input and output. This is especially bad if we have multi-stage pipeline, which keeps multiple copies of input for each table.

Prefetched embedding has the worst performance. By heuristically analyzing the current memory snapshot, we noticed 4~7x of extra input. 

This diff use a new formulas to calculate HBM usage:
- Multiple input are considered depending on the pipeline
- Output and input are added using max, since output tensor is not occupying extra HBM before a2a communication, at that time input is no longer used

For backward compatibility, we haven't roll out the change to ally user as it may incur extra failure of jobs. Instead, we set up different pipeline-awareness mode:
- None: Use old formulas regardless
- Prefetch-Only (default): Use new formulas if prefetch pipeline is on
- All: Use new formulas.

Differential Revision: D56444328