Simplify custom op naming for meta functionalization of sparse modules (#1974)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/1974

Registering custom ops for meta functionalization with ids can lead to hash collisions, resulting in wrong dimensions for a sparse module. This diff replaces custom op naming to just the dimension that is returned, alongside the module type, to ensure that the right dimensions are always returned and simplify the custom op naming logic significantly.

Reviewed By: TroyGarden

Differential Revision: D57108438

fbshipit-source-id: c547aaa9d29362124a8e8891dc301353ee6f9eac