One table per TBE mode will NOT separate multiple shards within the same table
Summary:
In the past, we assign UUID per each `ShardedEmbeddingTable` to ensure they're grouped into separate TBEs accordingly. However, we observed that it's possible for some tables (especially for VLE) to be have multiple shards of a table in the same rank, in this case, having them in different TBE will cause issue (see test plan).

This diff ensure multiple shards of a same table (sharing the same name) to be grouped into one TBE.

Differential Revision: D55671831