Fix dynamic embedding bugs that may lead to segmentation faults (#1947)
Summary:
Hi TorchRec Team,

We've recently been trying to use the dynamic embedding feature in torchrec contrib, but we  have encountered a few challenges. The process may result in a segmentation fault. After debugging, we've identified two potential problems.

1） fetch_notifications_ is not thread-safe.
2）Tensor data might be recycled during pushing operations.

For the first issue, the fetch_notifications_ variable is accessed in both the pull and sync fetch functions. As these functions are called in different threads, there might be a thread safety issue.

For the second issue, the tensor generated by concat in the push function is a temporary variable. The subsequent call to io.push is asynchronous and does not copy the data. Therefore, the data might have been recycled by the time it is executed asynchronously, leading to access to invalid data.

In light of these findings, we have undertaken some corrective measures to address these issues. We appreciate your attention to this Pull Request and eagerly await your valuable feedback.

Thank you for your time and consideration.

Pull Request resolved: https://github.com/pytorch/torchrec/pull/1947

Reviewed By: colin2328

Differential Revision: D57506220

Pulled By: PaulZhang12

fbshipit-source-id: f1522b855bdcb7f6bb1625d723b92ce612dc3228