Support flushing in StagedTrainPipeline
Summary:
Support ability to flush the current pipeline (i.e. finish processing all the inflight batches while pausing reading from dataloader), run a custom callback, and then resume again.

The trainer can check if the pipeline should be flushed, call `pipeline.set_flush(True)` to mark the pipeline to be flushed, and expect `on_flush_end` callback function to be invoked when it is completely flushed. I decided against naming the API `pipeline.flush()` as it might wrongly convey to users that pipeline will be immediately flushed upon invocation. Rather, we are marking the pipeline as being flushed so that it can internally handle calling the post-flush callback.

Reviewed By: zzzwen, murphymatt

Differential Revision: D57520129