Add imbalance stats for total perf, hbm and ddr (#2040)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2040

Add and log 4 ways to measure imbalance of generated sharding plan.

Suppose we have $k$ gpus. Let $s$ be a vector where the $i^{th}$ component is the total size currently allocated to the $i^{th}$ device.

First we normalize vector $s$ such that sum of elements in $p$ equals 1. We can view this as a probability distribution, and to get a measure of imbalance, we can measure its deviation from uniform distribution $p$ using one of the following ways:

- total variations
- total distance
- chi divergence
- KL divergence

Reviewed By: henrylhtsang

Differential Revision: D57465383

fbshipit-source-id: a22d59b17461dea46478b12beafe1b95c3331f62