separate out input dist tensors a2a out of embedding lookup stage to make it async (#2069)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2069

in the semi sync pipeline the start embedding lookup stage will call wait async on the input dist splits a2a then synchronously call wait on the tensors a2a. this will block the cpu waiting for the tensors a2a to finish. This diff follows the paradigm we have in train pipeline sparse dist which separates out the input dist into 2 stages so the a2as can be done asynchronously.

also in the embedding lookup stage we need to call wait on the tensor requests first so that the d2h sync that exists won't wait on work from previous module.

Reviewed By: dstaay-fb

Differential Revision: D58117193

fbshipit-source-id: 292f8b6b259fbdfe4f953994e94ae523487efcc6