add train_dlrm_data_parallel to show how to seperate dense optimizers for data_parallel tables and other dense (non embedding) parameters (#2084)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2084

how to optimize both data_parallel tables (non fused), model_parallel (fused) and dense params (non embeddings) is confusing to users.

This shows an example of how to do so.
We manually force one table (through parameter constraints) of DLRM example to be data_parallel.

Then, we set up optimizer on this table separately ( to be RWAdagrad dense optim). We keep the rest of th emodel as Adagrad

Reviewed By: henrylhtsang

Differential Revision: D58258207

fbshipit-source-id: 557adccdfefc98582da8dc8074287acf197118cf