Make train pipeline account for detached arg from torch.export serialization solution (#2086)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2086

With the torchrec serialization solution, previous modifications on the KJT might not be properly recorded. https://www.internalfb.com/phabricator/paste/view/P1399383958 is an example of a model where this bug surfaced.

This is due to the fact that in TorchRec serialization for the PT2 IR, the sparse arch is first compiled as part of the IR and then module swapped back to the original eager module say in the trainer. In order to do this, we preserve the module call signature of the sparse arch, making the inputs as KJTs instead of flattened values, lengths, weights.

However, in the train pipeline, this was not accounted for. If the input node was a KJT, only the values was looked at to see if it was modified, not the weights. This change now checks all the args for the KJT to see if they are a call_module (result of some other module output). If that is the case, there is no way this KJT is unmodified and should not be counted.

Differential Revision: D58272814