attempt to write specialized reshape code for 1d <-> 2d (#2075)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/2075

This diff contains torchrec changes that, together with a separate PyTorch PR, will ensure that uniform batch test with view conversions work in the existing torchrec tests.

I am deliberately not including the changes in test_pt2_multiprocess.py that I used to test locally on my devserver. I will let the torchrec team decide if they want those changes in this diff or they prefer to do it themselves in a separate diff.

Reviewed By: IvanKobzarev

Differential Revision: D58200241

fbshipit-source-id: b68e2cebce7e75b34a36f9ecf6ff546b3a1a3d8c