Fix TorchRec inference solution with updated torch.fx + TorchScript Solution (#2101)
Summary:
Previously, the TorchRec inference solution used torch.deploy, which isn't supported, and the solution itself was outdated and broken. This PR revamps the inference example with Torch.FX and TorchScript, which more accurately represents what is used in production for RecSys currently. Furthermore, the example was constructed with simplicity as the top priority, with fewer package requirements and steps to run. This is the first iteration of the inference example, with more to come!

Pull Request resolved: https://github.com/pytorch/torchrec/pull/2101

Differential Revision: D58478461

Pulled By: PaulZhang12